+++ MemoryManagement/GarbageCollection/Program.cs
@@ -2,3 +2,4 @@
 
 Task3 task3 = new Task3();
 task3.GenerateLargeNumberOfObjects();
+Console.ReadLine();
