+++ MemoryManagement/ValueAndReferenceTypes/Task1.cs
@@ -1,16 +1,20 @@
-﻿namespace ValueAndReferenceTypes
+﻿using ValueAndReferenceLibrary;
+
+namespace ValueAndReferenceTypes
 {
     /// <summary>
     /// Used to implement the task1
     /// </summary>
     public class Task1
     {
+        
+
         /// <summary>
         /// ExecuteTask1 the task1
         /// </summary>
         public void ExecuteTask1()
         {
-            ValueType valueType = new ValueType(10);
+            ValueAndReferenceLibrary.ValueType valueType = new ValueAndReferenceLibrary.ValueType(10);
             ReferenceType referenceType = new ReferenceType(10);
             Console.WriteLine($" Before value change for valueType :{valueType.Value}");
             Console.WriteLine($" Before value change for referenceType :{referenceType.Value}");
@@ -24,7 +28,7 @@
         /// </summary>
         /// <param name="valueType">struct as value type</param>
         /// <param name="referenceType">class as reference type</param>
-        private void changeValues(ValueType valueType, ReferenceType referenceType)
+        private void changeValues(ValueAndReferenceLibrary.ValueType valueType, ReferenceType referenceType)
         {
             valueType.Value = 20;
             referenceType.Value = 20;
